refactor: 게임이 진행중인 상태들을 추상 클래스 Running의 상속 관계로 묶어줌
