+++ src/main/java/blackjack/domain/state/Hit.java
@@ -1,29 +1,10 @@
 package blackjack.domain.state;
 
-import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
-import java.util.Objects;
 
-public class Hit implements State {
-
-    private final Cards cards;
+public class Hit extends Running {
 
     Hit(Cards cards) {
-        Objects.requireNonNull(cards, "[ERROR] 카드패는 null일 수 없습니다.");
-        this.cards = cards;
-    }
-
-    @Override
-    public State hit(Card card) {
-        Cards cards = this.cards.add(card);
-        if (cards.isBust()) {
-            return new Bust(cards);
-        }
-        return new Hit(cards);
-    }
-
-    @Override
-    public State stand() {
-        return new Stand(cards);
+        super(cards);
     }
 }
