+++ src/main/java/blackjack/domain/state/Running.java
@@ -1,0 +1,29 @@
+package blackjack.domain.state;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
+import java.util.Objects;
+
+public abstract class Running implements State {
+
+    private final Cards cards;
+
+    protected Running(Cards cards) {
+        Objects.requireNonNull(cards, "[ERROR] 카드패는 null일 수 없습니다.");
+        this.cards = cards;
+    }
+
+    @Override
+    public State hit(Card card) {
+        Cards cards = this.cards.add(card);
+        if (cards.isBust()) {
+            return new Bust(cards);
+        }
+        return new PlayerHit(cards);
+    }
+
+    @Override
+    public State stand() {
+        return new Stand(cards);
+    }
+}
